+++ R/mplm.R
@@ -24,10 +24,10 @@
 #' @export
 
 mplm <- function(data, dvar, mvar, pvar, 
-                 model = "W", 
-                 contrast = "first", 
-                 contrast_level = NA,
-                 contrast_slope = NA,
+                 model = c("W", "H-M", "B&L-B", "JW"),
+                 contrast = c("first", "preceding"),
+                 contrast_level = c(NA, "first", "preceding"),
+                 contrast_slope = c(NA, "first", "preceding"),
                  trend = TRUE, 
                  level = TRUE, 
                  slope = TRUE, 
@@ -35,6 +35,11 @@ mplm <- function(data, dvar, mvar, pvar,
                  update = NULL, 
                  na.action = na.omit, ...) {
  
+  model <- match.arg(model)
+  contrast <- match.arg(contrast)
+  contrast_level <- match.arg(contrast_level)
+  contrast_slope <- match.arg(contrast_slope)
+  
   if (is.na(contrast_level)) contrast_level <- contrast
   if (is.na(contrast_slope)) contrast_slope <- contrast
   
@@ -44,10 +49,14 @@ mplm <- function(data, dvar, mvar, pvar,
     model <- "B&L-B"
   }
   
-  start_check() %>%
-    check_in(model, c("H-M", "B&L-B", "W")) %>%
-    check_in(contrast, c("first", "preceding")) %>%
-    end_check()
+  check_args(
+    one_of(model, c("H-M", "B&L-B", "W")),
+    one_of(contrast, c("first", "preceding"))
+  )
+  #start_check() %>%
+  #  check_in(model, c("H-M", "B&L-B", "W")) %>%
+  #  check_in(contrast, c("first", "preceding")) %>%
+  #  end_check()
   
   # set attributes to arguments else set to defaults of scdf
   if (missing(dvar)) dvar <- scdf_attr(data, .opt$dv) else scdf_attr(data, .opt$dv) <- dvar